Better DMARC Compliance with reply-to working!
Outlook 2007 and earlier will still barf on this, but with MODERN Outlook and email clients, the Reply-To field is **respected** by the system when replying to a list.  This makes it change how it's displayed in Outlook, so that it should show 'original-sender@domain.tld on behalf of list@lists.domain.tld' rather than having to add a CC field with the original recipient.

DMARC in Windows and Microsoft implementations has been a little wonky this way, unfortunately...